requirements.txt
# Slack Intelligence - Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Slack SDK
slack-sdk==3.26.2

# OpenAI
openai>=1.50.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
httpx==0.28.1
python-multipart==0.0.6
rich==13.7.0

# Date/Time
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development (optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Exa Search
exa-py==2.0.1

# Jira Integration
atlassian-python-api==3.41.0

streamlit==1.28.0

# Scheduler
apscheduler==3.10.4

# Async HTTP (for Slack async client)
aiohttp==3.13.2

# Vector Database
pinecone>=3.0.0
